added tokens to required endpoints

+++ server/graphql/resolvers/pullRequests/deletePullRequest.js
@@ -7,8 +7,10 @@ const {
   deletePullRequestSuccess,
 } = require('./constants');
 
-const deletePullRequest = async ({ id }) => {
+const deletePullRequest = async ({ id }, { authError, userId }) => {
   try {
+    if (authError || !userId) throw new Error(authError);
+
     const result = await deletePullRequestQuery({ pullRequestId: id });
     await updateUserArray({
       column: 'pull_requests',
@@ -22,9 +24,10 @@ const deletePullRequest = async ({ id }) => {
       message: deletePullRequestSuccess,
     };
   } catch (error) {
+    const { message } = error;
     return {
       __typename: 'Error',
-      message: deletePullRequestError,
+      message: message || deletePullRequestError,
     };
   }
 };
